Parameterized gulpfile & organized project better
There is definitely a lot of room for improvement and for movement and
growth in this organization strategy. The purpose here was to build
things in a way that does not a require a manual modification of the
gulpfile every time a new file is added. However, this is definitely not
a target project structure.